Fix build in big endian
currently getting the following:
error: comparison of integer expressions of
different signedness: 'int' and 'long unsigned int'

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5392310
Auto-Submit: Milad Farazmand <mfarazma@redhat.com>
Commit-Queue: Milad Farazmand <mfarazma@redhat.com>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93039}

+++ src/compiler/turboshaft/operations.cc
@@ -1476,9 +1476,9 @@ template <size_t size>
 void PrintSimdValue(std::ostream& os, const uint8_t (&value)[size]) {
   os << "0x" << std::hex << std::setfill('0');
 #ifdef V8_TARGET_BIG_ENDIAN
-  for (int i = 0; i < size; i++) {
+  for (int i = 0; i < static_cast<int>(size); i++) {
 #else
-  for (int i = size - 1; i >= 0; i--) {
+  for (int i = static_cast<int>(size) - 1; i >= 0; i--) {
 #endif
     os << std::setw(2) << static_cast<int>(value[i]);
   }
